requirements.txt
flask
flask-restx
flask-cors
pillow